Moved ColorScale into /helpers
