+++ src/helpers/ColorScale.ts
